Mid stage commit
