+++ src/main/java/com/seifernet/wissen/configuration/OAuth2AuthorizationServerConfiguration.java
@@ -5,7 +5,6 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-/**
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.oauth2.config.annotation.configurers.ClientDetailsServiceConfigurer;
@@ -16,44 +15,30 @@ import org.springframework.security.oauth2.provider.ClientDetailsService;
 import org.springframework.security.oauth2.provider.token.DefaultTokenServices;
 import org.springframework.security.oauth2.provider.token.TokenStore;
 import org.springframework.security.oauth2.provider.token.store.InMemoryTokenStore;
-**/
-
-import com.seifernet.wissen.service.AccountClientDetailsService;
-import com.seifernet.wissen.service.AccountUserDetailsService;
 
 /**
  * OAuth2 authorization server configuration
  * 
- * @author Seifer (Cuauhtemoc Herrera)
- * @version 0.0.1
- *
+ * @author Seiferson (Cuauhtemoc Herrera)
  */
-//@Configuration
-//@EnableAuthorizationServer
-public class OAuth2AuthorizationServerConfiguration { /*extends AuthorizationServerConfigurerAdapter{
+@Configuration
+@EnableAuthorizationServer
+public class OAuth2AuthorizationServerConfiguration extends AuthorizationServerConfigurerAdapter{
 	
 	private TokenStore tokenStore = new InMemoryTokenStore();
-
+	
 	@Autowired
 	@Qualifier("authenticationManagerBean")
 	private AuthenticationManager authenticationManager;
-
-	@Autowired
-	private UserDetailsService userDetailsService;
 	
 	@Autowired
 	private ClientDetailsService clientDetailsService;
 	
 	@Autowired
 	private CustomProperties properties;
-
-	@Override
-	public void configure(AuthorizationServerEndpointsConfigurer endpoints) throws Exception {
-		endpoints
-			.tokenStore(this.tokenStore)
-			.authenticationManager(this.authenticationManager)
-			.userDetailsService(userDetailsService);
-	}
+	
+		@Autowired
+	private UserDetailsService userDetailsService;
 	
 	@Override
 	public void configure(ClientDetailsServiceConfigurer clients ) throws Exception {
@@ -67,7 +52,7 @@ public class OAuth2AuthorizationServerConfiguration { /*extends AuthorizationSer
 				.resourceIds("restservice")
 				.secret(properties.getMainClientAppSecret());
 	}
-
+	
 	@Bean
 	@Primary
 	public DefaultTokenServices tokenServices() {
@@ -76,5 +61,12 @@ public class OAuth2AuthorizationServerConfiguration { /*extends AuthorizationSer
 		tokenServices.setTokenStore(this.tokenStore);
 		return tokenServices;
 	}
-	*/
+	
+	@Override
+	public void configure(AuthorizationServerEndpointsConfigurer endpoints) throws Exception {
+		endpoints
+			.tokenStore(this.tokenStore)
+			.authenticationManager(this.authenticationManager)
+			.userDetailsService(this.userDetailsService);
+	}
 }
