+++ src/main/java/com/seifernet/wissen/configuration/OAuth2ResourceServerConfiguration.java
@@ -1,41 +1,32 @@
 package com.seifernet.wissen.configuration;
 
 import org.springframework.context.annotation.Configuration;
-/*
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
 import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
 import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
-*/
 
 /**
  * OAuth2 resource server configuration
  * 
- * @author Seifer (Cuauhtemoc Herrera)
- * @version 0.0.1
- *
+ * @author Seiferson (Cuauhtemoc Herrera)
  */
-//@Configuration
-//@EnableResourceServer
-public class OAuth2ResourceServerConfiguration {/*extends ResourceServerConfigurerAdapter{
+@Configuration
+@EnableResourceServer
+public class OAuth2ResourceServerConfiguration extends ResourceServerConfigurerAdapter{
 
 	@Override
 	public void configure(ResourceServerSecurityConfigurer resources) {
-		resources
-			.resourceId("restservice");
+		resources.resourceId("restservice");
 	}
 
 	@Override
 	public void configure(HttpSecurity http) throws Exception {
 		http
 			.authorizeRequests()
-				.antMatchers("/api/browser/*")
-					.permitAll()
-				.antMatchers("/api/browser")
-					.permitAll()
 				.antMatchers("/api")
 					.authenticated()
 				.antMatchers("/api/*")
 					.authenticated();
-	}*/
+	}
 }
