+++ src/main/java/com/seifernet/wissen/configuration/WebSecurityConfiguration.java
@@ -3,54 +3,36 @@ package com.seifernet.wissen.configuration;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-/*
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.encoding.ShaPasswordEncoder;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.web.util.matcher.AntPathRequestMatcher;*/
+
+import com.seifernet.wissen.service.CustomUserDetailsService;
 
 /**
- * Application security configuration
+ * Web security configuration
  * 
- * @author Seifer (Cuauhtemoc Herrera)
- * @version 0.0.1
- *
+ * @author Seiferson (Cuauhtemoc Herrera)
  */
-//@Configuration
-public class WebSecurityConfiguration { /*extends WebSecurityConfigurerAdapter {
-
+@Configuration
+public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
+	
 	@Override
 	protected void configure(HttpSecurity http) throws Exception {
 		http
 			.csrf()
 				.disable()
 			.authorizeRequests()
-				.antMatchers("/")
-					.authenticated()
-				.and()
-				.formLogin()
-					.loginPage("/login")
-					.permitAll()
-				.and()
-				.logout()
-					.logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
-					.logoutSuccessUrl("/")
+				.anyRequest()
 					.permitAll();
-    }
-
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth, UserDetailsService userDetailsService) throws Exception {
-    	auth
-    		.userDetailsService(userDetailsService)
-    		.passwordEncoder(new ShaPasswordEncoder(512));
-    }
-    
-    @Override
+	}
+	
+	@Autowired
+	private CustomUserDetailsService userDetailsService;
+	
+	@Override
 	@Bean
 	public AuthenticationManager authenticationManagerBean() throws Exception {
 		return super.authenticationManagerBean();
-	}*/
+	}
 }
