+++ src/main/java/com/seifernet/wissen/model/AccountGrantedAuthority.java
@@ -1,13 +1,13 @@
 package com.seifernet.wissen.model;
 
-//import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.GrantedAuthority;
 
 /**
+ * Model to save granted authorities as string on database
  * 
- * @author cuauhteh
- *
+ * @author Seiferson (Cuauhtemoc Herrera)
  */
-public class AccountGrantedAuthority {//implements GrantedAuthority{
+public class AccountGrantedAuthority implements GrantedAuthority{
 
 	private static final long serialVersionUID = 1L;
 	
@@ -17,9 +17,8 @@ public class AccountGrantedAuthority {//implements GrantedAuthority{
 		this.authority = authority;
 	}
 
-	//@Override
+	@Override
 	public String getAuthority() {
 		return authority;
 	}
-
 }
