+++ src/main/java/com/seifernet/wissen/repository/AccountRepository.java
@@ -1,0 +1,4 @@
+package com.seifernet.wissen.repository;
+
+public class AccountRepository extends MongoRepository<Author, String>{
+}
