+++ src/main/java/com/seifernet/wissen/util/URL.java
@@ -1,5 +1,10 @@
 package com.seifernet.wissen.util;
 
+/**
+ * URL definitions
+ * 
+ * @author Seiferson (Cuauhtemoc Herrera)
+ */
 public class URL {
 	public static final String NEW_USER = "/newuser";
 	public static final String INDEX = "/";
