+++ src/main/resources/application.properties
@@ -1,14 +1,17 @@
-#spring.data.mongodb.host: ds139198.mlab.com
-#spring.data.mongodb.host:localhost
-#spring.data.mongodb.username: seifer
-#spring.data.mongodb.password: s31f3rxx
-#spring.data.mongodb.port: 39198
-#spring.data.mongodb.port: 27017
-#spring.data.mongodb.database: wissen
+#General
 
 spring.data.rest.base-path=/api
 
+#Prod env
+
+#spring.data.mongodb.host = localhost
+#spring.data.mongodb.username = seifer
+#spring.data.mongodb.port = 27017
+#spring.data.mongodb.database = wissen
+#server.port = 80
+#wissen.main-client-app-id = 
+#wissen.main-client-app-secret = 
+
+#Testing env
 wissen.main-client-app-id=master
 wissen.main-client-app-secret=123456
-
-#server.port=80
